最长公共子序列问题

+++ IV Advanced Design and Analysis Techniques/Dynamic Programming/longestCommonSubsequence/longestCommonSubsequence/longestCommonSubsequence.h
@@ -14,7 +14,7 @@ template<typename T>
 void lcsLength(const vector<T>& _x, const vector<T>& _y, vector<vector<int32_t>>& _b, vector<vector<int32_t>>& _c);
 
 template<typename T>
-void printLcs(const vector<vector<int32_t>>& _b, const vector<T>& _x, size_t _i, size_t _j);
+void printLcs(const vector<vector<int32_t>>& _b, const vector<T>& _x, int _i, int _j);
 
 #endif // !_LONGEST_COMMON_SUBSEQUENCE_H_
